refactor to avoid importing functionality into shared types file
