+++ packages/back-end/src/events/base-events.ts
@@ -1,6 +1,6 @@
 import { ApiExperiment, ApiFeature } from "../../types/openapi";
-import { AuditableUserProperties } from "../services/users";
 import { NotificationEventPayload } from "./base-types";
+import { AuditableUserProperties } from "./events";
 
 // region User
