+++ packages/back-end/src/events/events.ts
@@ -1,0 +1,13 @@
+// region Audit user login events
+
+export type AuditableUserProperties = {
+  id: string;
+  email: string;
+  name: string;
+  device: string;
+  userAgent: string;
+  ip: string;
+  os: string;
+};
+
+// endregion Audit user login events
