+++ packages/back-end/src/services/users.ts
@@ -1,11 +1,12 @@
 import crypto from "crypto";
 import { promisify } from "util";
-import { Request } from "express";
 import uniqid from "uniqid";
+import { Request } from "express";
 import { UserDocument, UserModel } from "../models/UserModel";
+import { findOrganizationsByMemberId } from "../models/OrganizationModel";
 import { UserLoginNotificationEvent } from "../events/base-events";
 import { createEvent } from "../models/EventModel";
-import { findOrganizationsByMemberId } from "../models/OrganizationModel";
+import { AuditableUserProperties } from "../events/events";
 import { usingOpenId, validatePasswordFormat } from "./auth";
 
 const SALT_LEN = 16;
@@ -41,17 +42,58 @@ async function hash(password: string): Promise<string> {
   return salt + ":" + derivedKey.toString("hex");
 }
 
-// region Audit user login events
+export async function verifyPassword(
+  user: UserDocument,
+  password: string
+): Promise<boolean> {
+  if (!user.passwordHash) return false;
+  const [salt, key] = user.passwordHash.split(":");
+  const keyBuffer = Buffer.from(key, "hex");
+  const derivedKey = await (scrypt(
+    password,
+    salt,
+    HASH_LEN
+  ) as Promise<Buffer>);
+  return crypto.timingSafeEqual(keyBuffer, derivedKey);
+}
 
-export type AuditableUserProperties = {
-  id: string;
-  email: string;
-  name: string;
-  device: string;
-  userAgent: string;
-  ip: string;
-  os: string;
-};
+export async function updatePassword(userId: string, password: string) {
+  validatePasswordFormat(password);
+  const passwordHash = await hash(password);
+
+  await UserModel.updateOne(
+    {
+      id: userId,
+    },
+    {
+      $set: {
+        passwordHash,
+      },
+    }
+  );
+}
+
+export async function createUser(
+  name: string,
+  email: string,
+  password?: string,
+  verified: boolean = false
+) {
+  let passwordHash = "";
+
+  if (!usingOpenId()) {
+    password = validatePasswordFormat(password);
+    passwordHash = await hash(password);
+  }
+
+  return UserModel.create({
+    name,
+    email,
+    passwordHash,
+    id: uniqid("u_"),
+    verified,
+  });
+}
 
 /**
  * Some tracking properties exist on the request object
@@ -140,59 +182,4 @@ export async function trackLoginForUser({
   for (const organizationId of organizationIds) {
     await createEvent(organizationId, event);
   }
-}
-
-// endregion Audit user login events
-
-export async function verifyPassword(
-  user: UserDocument,
-  password: string
-): Promise<boolean> {
-  if (!user.passwordHash) return false;
-  const [salt, key] = user.passwordHash.split(":");
-  const keyBuffer = Buffer.from(key, "hex");
-  const derivedKey = await (scrypt(
-    password,
-    salt,
-    HASH_LEN
-  ) as Promise<Buffer>);
-  return crypto.timingSafeEqual(keyBuffer, derivedKey);
-}
-
-export async function updatePassword(userId: string, password: string) {
-  validatePasswordFormat(password);
-  const passwordHash = await hash(password);
-
-  await UserModel.updateOne(
-    {
-      id: userId,
-    },
-    {
-      $set: {
-        passwordHash,
-      },
-    }
-  );
-}
-
-export async function createUser(
-  name: string,
-  email: string,
-  password?: string,
-  verified: boolean = false
-) {
-  let passwordHash = "";
-
-  if (!usingOpenId()) {
-    password = validatePasswordFormat(password);
-    passwordHash = await hash(password);
-  }
-
-  return UserModel.create({
-    name,
-    email,
-    passwordHash,
-    id: uniqid("u_"),
-    verified,
-  });
 }
